crimson/osd: move start_pg_operation to pg_shard_manager
