+++ src/crimson/osd/osd_operation.h
@@ -177,9 +177,9 @@ protected:
   template <class OpT>
   friend class crimson::os::seastore::OperationProxyT;
 
-  // OSD::start_pg_operation needs access to enter_stage, we can make this
+  // PGShardManager::start_pg_operation needs access to enter_stage, we can make this
   // more sophisticated later on
-  friend class OSD;
+  friend class PGShardManager;
 };
 
 /**
