+++ src/crimson/osd/pg_shard_manager.h
@@ -84,8 +84,6 @@ public:
   FORWARD_TO_CORE(stop_pgs)
   FORWARD_CONST(get_pg_stats, get_pg_stats, core_state)
 
-  FORWARD_TO_CORE(get_or_create_pg)
-  FORWARD_TO_CORE(wait_for_pg)
   FORWARD_CONST(for_each_pg, for_each_pg, core_state)
   auto get_num_pgs() const { return core_state.pg_map.get_pgs().size(); }
 
@@ -97,6 +95,71 @@ public:
   template <typename F>
   auto with_pg(spg_t pgid, F &&f) {
     return std::invoke(std::forward<F>(f), core_state.get_pg(pgid));
+  }
+
+  template <typename T, typename... Args>
+  auto start_pg_operation(Args&&... args) {
+    auto op = local_state.registry.create_operation<T>(
+      std::forward<Args>(args)...);
+    auto &logger = crimson::get_logger(ceph_subsys_osd);
+    logger.debug("{}: starting {}", *op, __func__);
+    auto &opref = *op;
+
+    auto fut = opref.template enter_stage<>(
+      opref.get_connection_pipeline().await_active
+    ).then([this, &opref, &logger] {
+      logger.debug("{}: start_pg_operation in await_active stage", opref);
+      return core_state.osd_state.when_active();
+    }).then([&logger, &opref] {
+      logger.debug("{}: start_pg_operation active, entering await_map", opref);
+      return opref.template enter_stage<>(
+	opref.get_connection_pipeline().await_map);
+    }).then([this, &logger, &opref] {
+      logger.debug("{}: start_pg_operation await_map stage", opref);
+      using OSDMapBlockingEvent =
+	OSD_OSDMapGate::OSDMapBlocker::BlockingEvent;
+      return opref.template with_blocking_event<OSDMapBlockingEvent>(
+	[this, &opref](auto &&trigger) {
+	  std::ignore = this;
+	  return core_state.osdmap_gate.wait_for_map(
+	    std::move(trigger),
+	    opref.get_epoch(),
+	    &shard_services);
+	});
+    }).then([&logger, &opref](auto epoch) {
+      logger.debug("{}: got map {}, entering get_pg", opref, epoch);
+      return opref.template enter_stage<>(
+	opref.get_connection_pipeline().get_pg);
+    }).then([this, &logger, &opref] {
+      logger.debug("{}: in get_pg", opref);
+      if constexpr (T::can_create()) {
+	logger.debug("{}: can_create", opref);
+	return opref.template with_blocking_event<
+	  PGMap::PGCreationBlockingEvent
+	  >([this, &opref](auto &&trigger) {
+	    std::ignore = this; // avoid clang warning
+	    return core_state.get_or_create_pg(
+	      *this,
+	      shard_services,
+	      std::move(trigger),
+	      opref.get_pgid(), opref.get_epoch(),
+	      std::move(opref.get_create_info()));
+	  });
+      } else {
+	logger.debug("{}: !can_create", opref);
+	return opref.template with_blocking_event<
+	  PGMap::PGCreationBlockingEvent
+	  >([this, &opref](auto &&trigger) {
+	    std::ignore = this; // avoid clang warning
+	    return core_state.wait_for_pg(std::move(trigger), opref.get_pgid());
+	  });
+      }
+    }).then([this, &logger, &opref](Ref<PG> pgref) {
+      logger.debug("{}: have_pg", opref);
+      return opref.with_pg(get_shard_services(), pgref);
+    }).then([op] { /* Retain refcount on op until completion */ });
+
+    return std::make_pair(std::move(op), std::move(fut));
   }
 };
